chore: convert chrome_policy_check to TypeScript
